+++ test/unit/response.test.js
@@ -319,7 +319,7 @@ describe('Response', function () {
         it('should handle blank responses correctly', function () {
             var response = new Response();
             expect(response.size()).to.eql({
-                body: 0, header: 30, total: 30
+                body: 0, header: 32, total: 32
             });
         });
     });